Windows, affine bg improvements, cleanup in rendering

+++ spa/src/constants.rs
@@ -16,11 +16,11 @@ pub mod gba {
     pub const V_BLANK_RES: usize = 68;
 
     /// Width of bitmap in mode 5.
-    pub const SMALL_BITMAP_WIDTH: usize = 160;
+    pub const SMALL_BITMAP_WIDTH: u8 = 160;
     /// Height of bitmap in mode 5.
-    pub const SMALL_BITMAP_HEIGHT: usize = 128;
-    pub const SMALL_BITMAP_LEFT: usize = (H_RES - SMALL_BITMAP_WIDTH) / 2;
-    pub const SMALL_BITMAP_TOP: usize = (V_RES - SMALL_BITMAP_HEIGHT) / 2;
+    pub const SMALL_BITMAP_HEIGHT: u8 = 128;
+    pub const SMALL_BITMAP_LEFT: u8 = ((H_RES as u8) - SMALL_BITMAP_WIDTH) / 2;
+    pub const SMALL_BITMAP_TOP: u8 = ((V_RES as u8) - SMALL_BITMAP_HEIGHT) / 2;
 
     /// Cycles needed for a whole frame.
     pub const FRAME_CYCLES: usize = DOT_TIME * (H_RES + H_BLANK_RES) * (V_RES + V_BLANK_RES);
